admin dashboard

+++ local/regionalpoc/delete_regionalpoc.php
@@ -1,21 +1,31 @@
 <?php
 require_once('../../config.php');
 require_once($CFG->dirroot . '/user/lib.php');
+require_once($CFG->dirroot . '/local/regionalpoc/lib.php');
 require_login();
+local_regionalpoc_require_regional_manager();
 
 
-global $CFG, $DB;
+global $CFG, $DB, $OUTPUT, $USER;
     
 $id = optional_param('id', 0, PARAM_INT);
 // $userid = optional_param('userid', 0, PARAM_INT);
 $confirm=optional_param('confirm', 0, PARAM_INT);
-$usertype = optional_param('usertype', '', PARAM_TEXT);
+$usertype = 'arm';
 
+$armrecord = $DB->get_record('regionalpoc', ['userid' => $id, 'pocid' => $USER->id, 'usertype' => 'asstmanager']);
+if (!$armrecord) {
+    throw new required_capability_exception(context_system::instance(), 'local/pocschool:view', 'nopermissions', '');
+}
 
 if ($confirm) {
     if ($user = $DB->get_record('user', array('id' => $id))) {
         $deleted1 = user_delete_user($user);
         $deleted = $DB->delete_records('regionalpoc', array('userid' => $id));
+        $DB->delete_records('schoolassign', ['userid' => $id]);
+        if ($DB->get_manager()->table_exists('regionalpoc_arm_school')) {
+            $DB->delete_records('regionalpoc_arm_school', ['userid' => $id]);
+        }
         }
 
 
@@ -27,7 +37,7 @@ if ($confirm) {
 } else {
     echo $OUTPUT->header();
     echo $OUTPUT->confirm(get_string('deleteconfirm', 'local_regionalpoc'), 
-                         new moodle_url("$CFG->wwwroot/local/regionalpoc/delete_regionalpoc.php?confirm=1&id=$id?usertype=$usertype"), 
+                         new moodle_url('/local/regionalpoc/delete_regionalpoc.php', ['confirm' => 1, 'id' => $id, 'usertype' => $usertype]), 
                          new moodle_url("$CFG->wwwroot/local/regionalpoc/rm_arm_manage.php?usertype=$usertype"));
     echo $OUTPUT->footer();
 }
